add: Jumping and Idle Animation

+++ Entities/VelocityClass.cs
@@ -1,7 +1,7 @@
 public class VelocityClass
 {
     public int X { get; set; }
-    public int Y { get; set; }
+    public float Y { get; set; }
 
     public VelocityClass(int x, int y)
     {
